<rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"/> <?xml version="1.0" encoding="UTF-8"?>
The goal is to output an RDF/XML document.
XSLT has literal elements with the source code being closed under XML.
The RDF namespace is prefixed rdf, so these may be used later.
Template is a top-level element of XSLT 3.0.
Either the match or the name attribute is required for the XSLT template element.
The value of the match attribute is an XPath expression.
It selects the (only) root node of the input, i.xml .
This is a simple template.
Although it is terser to use the simplified syntax for a literal result element,
like is here now,
the goal requires a DOCTYPE declaration in the serialized output.
"Note that it is not possible, using a simplified stylesheet,
to request that the serialized output contains a DOCTYPE declaration."
The rdf:RDF root result will be the only static element determined by the goal.