Enable indenting.
"The xsl:output declaration is optional; if used, it MUST always appear as a top-level element with a stylesheet module.
"The default for the method attribute depends on the contents of the tree being serialized, and is chosen as follows."
The situation with the default method is determined by the first element child expanded QName not having the local part html.
The default is xml.
Paradoxically, output methods html and xhtml default to indenting, whereas xml does not.
"The default value is yes in the case of the html and xhtml output methods, no in the case of the xml output method."

While we are here, there are other xsl:output attributes that have already passed.

"The value of doctype-public must conform to the rules for a PubidLiteral (see [XML 1.0])."
This case was not tested for the PubidLiteral rules.
However, it does not seem worth pursuing given that these DTDs also have the PUBLIC slug within.

We already noted that we cannot set the XML document declaration encoding attribute to us-ascii.
"The default value is implementation-defined, but in the case of the xml and xhtml methods it MUST be either UTF-8 or UTF-16."
But is also says that there could be some XSLT processor that defaults to UTF-16.
If there were ever such a decision, then it raises the question of whether we should bother setting the attribute.
It would really depend on whether Prolog accepts the file.
Otherwise, without us-ascii, we need other tooling to set that.
Otherwise, with higher code-points, XML, RDF and Prolog accept those.