Switch rdf QNames to blank namespace.
Although the xsl:output undeclare-prefixes failed because of XML 1.0 serialization, there is another way to achieve the terse results that are prefered for keyboard typing. Simply omit the RDF namespace from within the stylesheet. XSLT 3.0 does offer XML Schema validation, so if that were used, it could interfer with this configuration. Without declaring xmlns:rdf, RDF does not have QName prefixes nor xmlns:rdf . So the literal xmlns is necessary for the rdf_load .