xsd ENTITY
Review of the test bed shows an additional entity.
The goal of matrix representation is an object literal.
Strings are the default type of RDF/XML.
Prolog recognizes integers without giving the xsd:string datatype,
but xsd:hexBinary is still parsed as a prolog string.
Using the datatype is conventional.
It specifies the data.
It doesn't save us from writing any prolog code.
So, this should be reverted.

Given that raster images are the goal of this transform,
one of the concerns is the prolog triple-store performance under high-resolution images.
For this reason the first approach is not to make the data totally parsed through color channel statements, which could multiply the total size.
Terser or more legible RDF/XML is a benefit by an alternate justification.
The assumption is still one object per coordinate pair,
which makes the structure of each matrix component like a set.