xsl:element attribute template
https://stackoverflow.com/questions/62286935/

Michael Kay of Saxonica answers on the subject of output entities.
He distinguishes between lexical XML and XDM,
which is that because of XDM, XSLT is agnostic of lexical XML.
The design argument has the reason character references.
From this, suppose that the lexical style solicits enhancement of the xml serializer.
&amp; is the reference in XML from both the stylesheet tree and to the result/output tree.
XDM ignores the stylesheet SGML <!ENTITY > "elements" , and is unable to produce them either.
The grammar of head(er) "elements" from SGML et al is pending.
From here, the performance of the streaming feature is still a resolution scale issue that makes XSLT an optimation:
both computationally and wrt the expressitivity of XML serialization.
Technically, the long URLs work for SWI-Prolog, and are neither an issue for gzip nor hypothesized triple-store indexing.
But the keyboard overhead of RDF/XML for projects presents resistance in the editor.
XSLT would also be a consideration for practical XML fragments transforms.
Without lexical style controls, one must make trade-offs or testing scripts.
So if there is a way with disable-output-escaping, then it would be good to iron that out.
ATM, the double sense of ampersand has made it questionable whether xsl:text or xsl:value-of can set an attribute value.