<xsl:variable static="yes" select="'http://www.w3.org/1999/02/22-... />
The intent of rdf/ns is certainly a constant immutable or static value.
The compiler gives an error with static="yes" and the prior sequence constructor.
So, in the first place, this commit really solves an alternative XPath expression by including apostrophies to make the string literal.
It saves a few source characters because the end-tag can be omitted.
But the code runs with and without static="yes" .
Spec 9.7 "Static Expressions" describes static analysis, static context, and static expression.
"These constaints are designed to ensure that the expresssion can be evaluated at the earliest possible stage of stylesheet processing, without any dependency on information contained in the stylesheet itself or in any source document."
Intra-source dependency apparently means not referring to other nodes, but referring to one's own leaves is acceptable.
This is why there cannot be a sequence constructor.
Alternatively again, xsl:variable might accept a certain namespace type which casts the string, perhaps performing like the normalize-space funcation.
So in that sense, if static="yes" only declares intent but does not change neither function nor output, then why are we keyboarding it?
Likewise, the xmlns:xsl and EQN prefixed tags eat up a lot of characters just so the processor doesn't pollute the output with <_1 xmlns="w3:xslt3...
xsd:langString: formal language tags are the way in which we parse, accept, decompose quotational encodings.
The HTML color hex codes make much more sense and readibility as literals, but are related by a mini-parser language to the more structured rgb integers.
This is a matrix raster.
An image.
If you can't view it, save it, rename it, then it is no good.
In the corner of that, is programming the image.
Reviewing "Metamathmatics" by S. C. Kleene, on enumeration and the structure of numbers: natural, integer, quotients, shows that these representations can be reduced to natural indexing.
So the format is really hyperdigraph translation format with datatype literals that needs lots of decisions to recover others' decisions.