<element namespace=""
Corrects the property namespace.
From the XSLT3 Spec, "If the string [of the namespace attribute value] is zero-length, then the expanded QName of the element has a null namespace URI."
This style will succeed for this project because xsl:element will only be invoked once or twice (lexically).
That usage is total less characters than the xsl: EQName.
However, the reading indicates that why we got here is because of ignoring the default namespace, apparently inherited from the stylesheet tree.
The xsl:disable-output-escaping attributes are required (for result entities) and accidental to any namespace.
xsl:element doesn't know they are in result tree as ancestors.
The null namespace URI has the intuition of relativizing the membership property attribute rdf:_1 upon external parse of the serialized result.
The external EQName doesn't exist in the result tree.
But RDF/XML is the only option for correct RDF namespace "unification" with Prolog namespace, otherwise the forbidden subject membership property _1 in Turtle is an atom of the expanded rdf:_1 URI but not equal to the built-in rdf namespace, that is, different namespaces.
Stardog for example, has an entity limit which practically only arises with the SWI RDF/XML.
Google's Tensor Chip just released to consumers in the Pixel 6 Phone is NN/arithmetic(float)-manifold.